+++ programs/Xserver/Xext/mbuf.c
@@ -1,4 +1,4 @@
-/* $XFree86: xc/programs/Xserver/Xext/mbuf.c,v 3.3 1997/01/18 06:52:58 dawes Exp $ */
+/* $XFree86: xc/programs/Xserver/Xext/mbuf.c,v 3.4 1998/10/04 09:36:44 dawes Exp $ */
 /************************************************************
 
 Copyright 1989, 1998  The Open Group
@@ -433,8 +433,8 @@ SetupBackgroundPainter (pWin, pGC)
     case BackgroundPixmap:
 	gcvalues[0] = (pointer) FillTiled;
 	gcvalues[1] = (pointer) background.pixmap;
-	gcvalues[2] = (pointer) ts_x_origin;
-	gcvalues[3] = (pointer) ts_y_origin;
+	gcvalues[2] = (pointer)(long) ts_x_origin;
+	gcvalues[3] = (pointer)(long) ts_y_origin;
 	gcmask = GCFillStyle|GCTile|GCTileStipXOrigin|GCTileStipYOrigin;
 	break;
